// src/graphql/generated.ts
export type GetBoardByIdQuery = { __typename?: 'query_root', boards_by_pk?: { __typename?: 'boards', id: any, name: string, columns: Array<{ __typename?: 'columns', id: any, name: string, order_index: any, cards: Array<{ __typename?: 'cards', id: any, title?: string | null, description?: string | null, assignee?: string | null, order_index: any }> }> } | null };